feat(ArtyHub-Mislata/sprints#36): Añadidos los tests menos los de controller

+++ src/main/java/es/artyhub/banco_back/domain/repository/MovimientoBancarioRepository.java
@@ -1,12 +1,13 @@
 package es.artyhub.banco_back.domain.repository;
 
+import java.math.BigDecimal;
 import java.util.List;
 
 import es.artyhub.banco_back.domain.model.MovimientoBancario;
 
 public interface MovimientoBancarioRepository {
     MovimientoBancario findById(Long id);
-    MovimientoBancario findByImporte(Long importe);
+    MovimientoBancario findByImporte(BigDecimal importe);
     MovimientoBancario findByConcepto(String concepto);
     List<MovimientoBancario> findByCuentaId(Long cuenta_id);
     List<MovimientoBancario> findAll();
