+++ src/main/java/es/artyhub/banco_back/domain/service/impl/MovimientoBancarioServiceImpl.java
@@ -1,5 +1,6 @@
 package es.artyhub.banco_back.domain.service.impl;
 
+import java.math.BigDecimal;
 import java.util.List;
 import es.artyhub.banco_back.domain.exception.ResourceNotFoundException;
 import es.artyhub.banco_back.domain.exception.ValidationException;
@@ -30,7 +31,7 @@ public class MovimientoBancarioServiceImpl implements MovimientoBancarioService 
     }
 
     @Override
-    public MovimientoBancario findByImporte(Long importe) {
+    public MovimientoBancario findByImporte(BigDecimal importe) {
         
         if(importe == null) {
             throw new ValidationException("Importe no valido");
